✨ refactor: rename Navbar file and move Particles component to correct directory
